Re-organize partner modules
