+++ admin_apps/app.py
@@ -69,7 +69,8 @@ if __name__ == "__main__":
         """
 
         # Direct to specific page based instead of default onboarding if user comes from successful partner setup
-        if st.session_state.get('partner_setup', False):
+        if (st.session_state.get("partner_setup", False) 
+            and st.session_state.get("partner_tool", None) == 'looker'):
             builder.show()
         st.markdown(
             """
